add constants for user interaction

+++ src/events/interactionCreate.ts
@@ -2,6 +2,7 @@ import { Message } from 'discord.js';
 import commands from '../commands'
 import { Command } from '../types'
 import { EditReply, event, Reply } from '../utils'
+import { idUser, idMember, idGuild } from '../commands/debug/constants';
 
 const allCommands = commands.map(({ commands }) => commands).flat()
 const allCommandsMap = new Map<string, Command>(
@@ -46,7 +47,20 @@ export default event('interactionCreate', async (
     if (interaction.isButton())
     {
         if(interaction.customId === 'delete'){
-            return await interaction.message.delete();
+            const user = client.guilds.cache.get(interaction.guild?.id!);
+            const member = user?.members.cache.get(interaction.user.id);
+            console.log(idMember[0]);
+            for(let i = 0; i < idMember.length; i++) {
+                const tagName = client.users.cache.get(idMember[i].id);
+                if(idMember[i] == member){
+                    idMember.splice(i - 1);
+                    return await interaction.message.delete();
+                }
+                else return interaction.reply({
+                    ephemeral: true,
+                    content: `Pesan hanya bisa dihapus oleh ${tagName?.username}`
+                });
+            }
         }
     }
